Refactor test suite structure and update namespaces
The commit renames and restructures test files, shifting them from the `Blueprint` namespace to more specific and relevant namespaces. This includes updating of `use` statements and class extensions accordingly. It also relocates certain utility functions to a central `Tools` class for shared use.
